+++ test/choice.js
@@ -2,8 +2,7 @@
 var assert = require("assert"),
 	state = require("../lib/node/state");
 
-var oldLogger = state.logger;
-state.setLogger(console);
+var oldLogger = state.setLogger(console);
 
 // this test overrides the default implementation of the random selector for choices as we're not looking to test the randomness of hte numbers, but the application of them to choose different transtiions therefore we need to turn the non-deterministic into something deterministic
 var nextRand = 0;
@@ -50,9 +49,7 @@ describe("test/choice.js", function () {
 	});
 
 	describe("With an non-random distribution, each path is called equally", function () {
-		var oldRandom = state.random;
-
-		state.setRandom(randRobin);
+		var oldRandom = state.setRandom(randRobin);
 
 		var instance2 = new state.DictionaryInstance("instance2");
 		instance2.path1 = 0;