+++ test/unit_model.js
@@ -1,8 +1,7 @@
 var assert = require("assert"),
 	state = require("../lib/node/state");
 
-var oldLogger = state.logger;
-state.setLogger(console);
+var oldLogger = state.setLogger(console);
 
 var model = new state.StateMachine("unit_model");
 var initial = new state.PseudoState("initial", model);